# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Name",
    "version": "1.0.0",
    "description": "API documentation for my project"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/employees": {
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "department",
                  "role",
                  "companyId"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "department": {
                    "type": "string",
                    "example": "Engineering"
                  },
                  "role": {
                    "type": "string",
                    "example": "Software Developer"
                  },
                  "companyId": {
                    "type": "integer",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Employee created successfully!"
          },
          "400": {
            "description": "one of the fields is not provided!"
          },
          "401": {
            "description": "not logged in / invalid or expired token"
          },
          "403": {
            "description": "Unauthorized user!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all employees",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Employees retrieved successfully!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "summary": "Get a single employee by its ID",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee retrieved successfully!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update an employee, at least one field is required",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully!"
          },
          "400": {
            "description": "Atleast one field should be privided!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee by its ID",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employee-accounts": {
      "post": {
        "summary": "Create a new employee account",
        "tags": [
          "Employee Accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "employeeId",
                  "username",
                  "password"
                ],
                "properties": {
                  "employeeId": {
                    "type": "integer",
                    "example": 3
                  },
                  "username": {
                    "type": "string",
                    "example": "jdoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "secretPassword123"
                  },
                  "permission": {
                    "type": "string",
                    "example": "READ"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully!"
          },
          "400": {
            "description": "employeeId, username or password is not provided!"
          },
          "401": {
            "description": "not logged in / invalid or expired token"
          },
          "403": {
            "description": "Unauthorized user!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all employee accounts",
        "tags": [
          "Employee Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Employee accounts retrieved successfully!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/employee-accounts/login": {
      "post": {
        "summary": "Log an employee in and receive a JWT",
        "tags": [
          "Employee Accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "jdoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "secretPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully"
          },
          "400": {
            "description": "user is not found, please sign up / login credintials are invaild"
          },
          "500": {
            "description": "username or password are incorrect, please try again!"
          }
        }
      }
    },
    "/employee-accounts/{id}": {
      "get": {
        "summary": "Get a single employee account by its ID",
        "tags": [
          "Employee Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee account retrieved successfully!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update an employee account, at least one field is required",
        "tags": [
          "Employee Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully!"
          },
          "400": {
            "description": "Atleast one field should be privided!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee account by its ID",
        "tags": [
          "Employee Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/companies": {
      "post": {
        "summary": "Create a new company",
        "tags": [
          "Companies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "industry"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Acme Inc."
                  },
                  "industry": {
                    "type": "string",
                    "example": "Technology"
                  },
                  "location": {
                    "type": "string",
                    "example": "New York"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Company created successfully"
          },
          "400": {
            "description": "name and industry are required"
          },
          "401": {
            "description": "not logged in / invalid or expired token"
          },
          "403": {
            "description": "Unauthorized user!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all companies",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Companies retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "summary": "Get a single company by its ID",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update a company, at least one field is required",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "industry": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully"
          },
          "400": {
            "description": "Atleast one field should be privided!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a company by its ID",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company deleted successfully from records!"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}